pyproject.toml
[tool.poetry]
name = "budget_assistant"
version = "0.1.0"
description = ""
authors = ["Roman <dubrovskijrd@yandex.ru>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.18"
aiohttp = "^3.8.4"
gunicorn = "^20.1.0"
asyncpg = "^0.27.0"
msgspec = "^0.16.0"
python-dotenv = "^1.0.0"
aiohttp-cors = "^0.7.0"
mashumaro = "^3.8.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.277"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
